requirements.txt
pip~=23.0.1
build~=1.2.2.post1
wheel~=0.37.0
Brotli~=1.1.0
tornado~=6.4.1
setuptools~=58.0.4
future~=0.18.2
python-dotenv~=1.0.1
langchain~=0.3.3
streamlit~=1.40.2
langchain_openai~=0.2.11
langchain-community~=0.3.10
pypdf~=5.1.0